don't dump anything to file if output_dir is not specified.

+++ droidbot/device.py
@@ -40,10 +40,9 @@ class Device(object):
         self.ro_secure = None
 
         self.output_dir = output_dir
-        if self.output_dir is None:
-            self.output_dir = os.path.abspath("droidbot_out")
-        if not os.path.exists(self.output_dir):
-            os.mkdir(self.output_dir)
+        if output_dir is not None:
+            if not os.path.isdir(output_dir):
+                os.mkdir(output_dir)
 
         self.use_hierarchy_viewer = use_hierarchy_viewer
         self.grant_perm = grant_perm
@@ -66,7 +65,7 @@ class Device(object):
         self.get_ro_secure()
         self.get_ro_debuggable()
         self.get_display_info()
-        self.logcat = self.redirect_logcat()
+        self.logcat = self.redirect_logcat(self.output_dir)
         from state_monitor import StateMonitor
         self.state_monitor = StateMonitor(device=self)
         self.state_monitor.start()
@@ -77,7 +76,7 @@ class Device(object):
 
     def redirect_logcat(self, output_dir=None):
         if output_dir is None:
-            output_dir = self.output_dir
+            return None
         logcat_file = open("%s/logcat.log" % output_dir, "w")
         import subprocess
         subprocess.check_call(["adb", "-s", self.serial, "logcat", "-c"])
@@ -176,12 +175,15 @@ class Device(object):
             self.monkeyrunner.disconnect()
         if self.view_client:
             self.view_client.disconnect()
-        self.logcat.terminate()
+        if self.logcat:
+            self.logcat.terminate()
         self.state_monitor.stop()
-        temp_dir = os.path.join(self.output_dir, "temp")
-        if os.path.exists(temp_dir):
-            import shutil
-            shutil.rmtree(temp_dir)
+
+        if self.output_dir is not None:
+            temp_dir = os.path.join(self.output_dir, "temp")
+            if os.path.exists(temp_dir):
+                import shutil
+                shutil.rmtree(temp_dir)
 
     def get_telnet(self):
         """
@@ -193,7 +195,7 @@ class Device(object):
             try:
                 self.telnet = TelnetConsole(self)
             except Exception:
-                self.logger.warning("Cannot connect to telnet. You may not be able to simulate phonecalls and locations.")
+                self.logger.info("Telnet not connected. If you want to enable telnet, please use an emulator.")
         return self.telnet
 
     def get_adb(self):
@@ -597,11 +599,17 @@ class Device(object):
         install_cmd.append(app.app_path)
         subprocess.check_call(install_cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
 
-        package_info_file_name = "%s/dumpsys_package_%s.txt" % (self.output_dir, app.get_package_name())
-        package_info_file = open(package_info_file_name, "w")
-        subprocess.check_call(["adb", "-s", self.serial, "shell",
-                               "dumpsys", "package", app.get_package_name()], stdout=package_info_file)
-        package_info_file.close()
+        if self.output_dir is not None:
+            package_info_file_name = "%s/dumpsys_package_%s.txt" % (self.output_dir, app.get_package_name())
+            package_info_file = open(package_info_file_name, "w")
+        else:
+            package_info_file = subprocess.PIPE
+
+        subprocess.check_call(["adb", "-s", self.serial, "shell", "dumpsys", "package",
+                               app.get_package_name()], stdout=package_info_file)
+
+        if isinstance(package_info_file, file):
+            package_info_file.close()
 
     def uninstall_app(self, app):
         assert isinstance(app, App)
@@ -664,6 +672,9 @@ class Device(object):
         # except IOError as e:
         #     self.logger.warning("exception in take_screenshot: %s" % e)
         # return image
+        if self.output_dir is None:
+            return None
+
         from datetime import datetime
         tag = datetime.now().strftime("%Y-%m-%d_%H%M%S")
 
@@ -803,7 +814,10 @@ class DeviceState(object):
     def save2dir(self, output_dir=None):
         try:
             if output_dir is None:
-                output_dir = os.path.join(self.device.output_dir, "states")
+                if self.device.output_dir is None:
+                    return
+                else:
+                    output_dir = os.path.join(self.device.output_dir, "states")
             if not os.path.exists(output_dir):
                 os.mkdir(output_dir)
             state_json_file_path = "%s/state_%s.json" % (output_dir, self.tag)
